+++ src/lib.rs
@@ -63,6 +63,6 @@ pub use theme::Theme;
 pub use widget::{PopupKind, PopupRequest, Widget};
 pub use widgets::{
     Bevel, Button, Checkbox, Column, Container, Dialog, DialogIcon, Dropdown, FileDialog,
-    FileFilter, Image, Label, List, ListIcon, ListItem, Menu, MenuBar, MenuItem, Modal,
+    FileFilter, FocusLabel, Image, Label, List, ListIcon, ListItem, Menu, MenuBar, MenuItem, Modal,
     Orientation, ProgressBar, Row, SCROLLBAR_THICKNESS, ScrollBar, Slider, TextEditor, TextInput,
 };